backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0  # Required for EmailStr validation

# Database
# psycopg2-binary==2.9.9  # Skipped - incompatible with Python 3.14, using Supabase client instead
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # Required for Alembic migrations

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Cryptography (for smart meter signatures)
cryptography==42.0.5

# Hedera SDK
hedera-sdk-py==2.21.0  # Official Hedera SDK for Python (wrapper around Java SDK)

# Google Cloud Vision (OCR)
google-cloud-vision==3.12.1  # Updated for Python 3.14 compatibility
protobuf>=6.33.5  # Required for Python 3.14

# IPFS
ipfshttpclient==0.8.0a2

# Exchange Rate APIs
requests==2.31.0

# Image Processing
pillow==12.1.1  # Updated for Python 3.14 compatibility
numpy>=2.4.2  # For fraud detection (ELA analysis)

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1  # For Mirror Node async HTTP requests

# Email
sendgrid==6.11.0

# PDF Generation
reportlab==4.0.9
qrcode[pil]==8.0  # QR code generation for receipts

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
